+++ FDW/include/D3DFramework/D3DFramework/GraphicUtilites/CommandQueue.h
@@ -11,7 +11,7 @@ namespace FD3DW
 	{
 
 	public:
-		static std::unique_ptr<CommandQueue> CreateQueue(ID3D12Device* pDevice, const D3D12_COMMAND_LIST_TYPE type, const D3D12_COMMAND_QUEUE_FLAGS flags = D3D12_COMMAND_QUEUE_FLAG_NONE, size_t priority = 0, size_t nodeMask = 0);
+		static std::unique_ptr<CommandQueue> CreateQueue(ID3D12Device* pDevice, const D3D12_COMMAND_LIST_TYPE type, const D3D12_COMMAND_QUEUE_FLAGS flags = D3D12_COMMAND_QUEUE_FLAG_NONE, INT priority = 0, UINT nodeMask = 0);
 
 	public:
